build: move `bin` and `lib` to root and remove build process for CRS itself
