+++ .eslintignore
@@ -1,5 +1,3 @@
 /coverage
-/es
-/lib
 /node_modules
 /templates
