+++ .gitignore
@@ -1,7 +1,5 @@
 /.nyc_output
 /coverage
-/es
-/lib
 /node_modules
 npm-debug.log*
 .DS_Store
