+++ bin/commands/script.js
