+++ bin/commands/scripts/build.js
