+++ bin/commands/scripts/build_styleguide.js
