+++ bin/commands/scripts/build_watch.js
