+++ bin/commands/scripts/clean.js
