+++ bin/commands/scripts/eslint.js
