+++ bin/commands/scripts/prepublishOnly.js
