+++ bin/commands/scripts/start.js
