+++ bin/commands/scripts/test.js
