+++ bin/commands/scripts/test_coverage.js
