+++ bin/commands/scripts/test_update.js
