+++ bin/commands/scripts/test_watch.js
