+++ bin/util/executables.js
