+++ jest.config.js
@@ -1,10 +1,12 @@
-const { createJestConfig } = require('./src');
+const { createJestConfig } = require('./lib');
 
 const config = createJestConfig();
 
 module.exports = {
     ...config,
-    // TODO: test the CLI properly
-    coveragePathIgnorePatterns: ['<rootDir>/src/bin/'],
-    testPathIgnorePatterns: [...config.testPathIgnorePatterns, '<rootDir>/templates/'],
+    // TODO: test the CLI properly]
+    collectCoverageFrom: ['lib/**/*.js', 'lib/**/*.jsx'],
+    coveragePathIgnorePatterns: ['<rootDir>/bin/'],
+    modulePathIgnorePatterns: [],
+    testPathIgnorePatterns: ['/node_modules/', '<rootDir>/templates/'],
 };
