+++ lib/configs/__tests__/createJestConfig.test.js
