+++ lib/configs/__tests__/createStyleguideConfig.test.js
