+++ lib/configs/createStyleguideConfig.js
