+++ lib/index.js
@@ -1,0 +1,7 @@
+const createJestConfig = require('./configs/createJestConfig');
+const createStyleguideConfig = require('./configs/createStyleguideConfig');
+
+module.exports = {
+    createJestConfig,
+    createStyleguideConfig,
+};
