+++ package.json
@@ -9,24 +9,23 @@
     "url": "https://github.com/zillow/create-react-styleguide/issues"
   },
   "bin": {
-    "create-react-styleguide": "./lib/bin/create-react-styleguide.js"
+    "create-react-styleguide": "./bin/index.js"
   },
   "main": "./lib/index.js",
   "scripts": {
-    "build": "src/bin/create-react-styleguide.js script build",
-    "build:watch": "src/bin/create-react-styleguide.js script build:watch",
-    "clean": "src/bin/create-react-styleguide.js script clean",
-    "eslint": "src/bin/create-react-styleguide.js script eslint",
-    "eslint:fix": "src/bin/create-react-styleguide.js script eslint:fix",
-    "test": "src/bin/create-react-styleguide.js script test",
-    "test:coverage": "src/bin/create-react-styleguide.js script test:coverage",
-    "test:update": "src/bin/create-react-styleguide.js script test:update",
-    "test:watch": "src/bin/create-react-styleguide.js script test:watch",
-    "prepublishOnly": "src/bin/create-react-styleguide.js script prepublishOnly",
+    "clean": "./bin/index.js script clean",
+    "eslint": "./bin/index.js script eslint",
+    "eslint:fix": "./bin/index.js script eslint:fix",
+    "test": "./bin/index.js script test",
+    "test:coverage": "./bin/index.js script test:coverage",
+    "test:update": "./bin/index.js script test:update",
+    "test:watch": "./bin/index.js script test:watch",
+    "prepublishOnly": "./bin/index.js script prepublishOnly",
     "release": "standard-version",
     "prepare": "husky install"
   },
   "files": [
+    "bin",
     "lib",
     "templates"
   ],
