--- src/index.js
@@ -1,7 +1,0 @@
-const createStyleguideConfig = require('./js/createStyleguideConfig');
-const createJestConfig = require('./js/createJestConfig');
-
-module.exports = {
-    createStyleguideConfig,
-    createJestConfig,
-};
